Mvn Build

+++ Backend/app/src/main/java/com/cetcollegefinder/app/controllers/UserController.java
@@ -10,12 +10,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
 
 //operations
 //GET /colleges
@@ -38,7 +36,7 @@ public class UserController {
     }
 
     @GetMapping("/userbyid/{id}")
-    public UserDTO findUserById(@PathVariable int id) {
+    public UserDTO findUserById(@PathVariable Long id) {
         return service.getUserById(id);
     }
 
@@ -63,7 +61,7 @@ public class UserController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public String deleteUser(@PathVariable int id) {
+    public String deleteUser(@PathVariable Long id) {
         return service.deleteUser(id);
     }
